Hide Westopolis until update
Even if it's 1 round, we cannot let this map slide in to be honest. most defense points needs a reestructure and depending on checkpoints tps is not good.

This is also to avoid new players to have wrong impresions about ze2 and the maps.

So hidden until ByPigs fixes it. I do also this because he's taking long already for many ze2 updates.