Nerf Flame Ring
This single item can single carry a whole round along with red ring considering his range and effect it gives to zombies.

I think reducing the ammo, increasing the reload by +1 second, and half the effect time is the best choice for an item like this.

Still gives enough time for the survivor to hold the zombies and escape. so it's not too far from it's purpose.